# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  username: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  name: string;
  username: string;
}

export interface AuthMeResponse {
  user: AuthUser | null;
}

export interface SocialLink {
  platform: string;
  url: string;
  label: string;
}

export interface LandingPhoto {
  objectPath: string;
  /** @nullable */
  caption?: string | null;
}

export interface LandingPage {
  id: number;
  userId: number;
  username: string;
  template: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  appFileObjectPath?: string | null;
  /** @nullable */
  tutorialVideoObjectPath?: string | null;
  /** @nullable */
  tutorialVideoUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  photos: LandingPhoto[];
  socialLinks: SocialLink[];
}

export interface UpdateLandingBody {
  template?: string;
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  appFileObjectPath?: string | null;
  /** @nullable */
  tutorialVideoObjectPath?: string | null;
  /** @nullable */
  tutorialVideoUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  photos?: LandingPhoto[];
  socialLinks?: SocialLink[];
}

export type GalleryItemType =
  (typeof GalleryItemType)[keyof typeof GalleryItemType];

export const GalleryItemType = {
  photo: "photo",
  video: "video",
} as const;

export interface GalleryItem {
  id: string;
  objectPath: string;
  type: GalleryItemType;
  /** @nullable */
  caption?: string | null;
  order: number;
}

export interface Gallery {
  id: number;
  userId: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  items: GalleryItem[];
}

export interface UpdateGalleryBody {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  items?: GalleryItem[];
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}