# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: pages
  - name: storage
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Create account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
        "409":
          description: Email or username already taken

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session
      responses:
        "200":
          description: Current user or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMeResponse"

  /my/landing:
    get:
      operationId: getMyLanding
      tags: [pages]
      summary: Get my landing page
      responses:
        "200":
          description: Landing page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandingPage"
        "401":
          description: Unauthorized
    put:
      operationId: updateMyLanding
      tags: [pages]
      summary: Update my landing page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLandingBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandingPage"
        "401":
          description: Unauthorized

  /my/gallery:
    get:
      operationId: getMyGallery
      tags: [pages]
      summary: Get my gallery
      responses:
        "200":
          description: Gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "401":
          description: Unauthorized
    put:
      operationId: updateMyGallery
      tags: [pages]
      summary: Update my gallery
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGalleryBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "401":
          description: Unauthorized

  /pages/{username}:
    get:
      operationId: getPublicLanding
      tags: [pages]
      summary: Get public landing page by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Landing page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LandingPage"
        "404":
          description: Not found

  /pages/{username}/gallery:
    get:
      operationId: getPublicGallery
      tags: [pages]
      summary: Get public gallery by username
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gallery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gallery"
        "404":
          description: Not found

  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request presigned upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestUploadUrlBody"
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestUploadUrlResponse"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        username:
          type: string
      required: [email, password, name, username]

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        username:
          type: string
      required: [id, email, name, username]

    AuthMeResponse:
      type: object
      properties:
        user:
          oneOf:
            - $ref: "#/components/schemas/AuthUser"
            - type: "null"
      required: [user]

    SocialLink:
      type: object
      properties:
        platform:
          type: string
        url:
          type: string
        label:
          type: string
      required: [platform, url, label]

    LandingPhoto:
      type: object
      properties:
        objectPath:
          type: string
        caption:
          type: ["string", "null"]
      required: [objectPath]

    LandingPage:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        template:
          type: string
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        appFileObjectPath:
          type: ["string", "null"]
        tutorialVideoObjectPath:
          type: ["string", "null"]
        tutorialVideoUrl:
          type: ["string", "null"]
        buttonText:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/LandingPhoto"
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"
      required: [id, userId, username, template, title, photos, socialLinks]

    UpdateLandingBody:
      type: object
      properties:
        template:
          type: string
        title:
          type: string
        subtitle:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        appFileObjectPath:
          type: ["string", "null"]
        tutorialVideoObjectPath:
          type: ["string", "null"]
        tutorialVideoUrl:
          type: ["string", "null"]
        buttonText:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/LandingPhoto"
        socialLinks:
          type: array
          items:
            $ref: "#/components/schemas/SocialLink"

    GalleryItem:
      type: object
      properties:
        id:
          type: string
        objectPath:
          type: string
        type:
          type: string
          enum: [photo, video]
        caption:
          type: ["string", "null"]
        order:
          type: integer
      required: [id, objectPath, type, order]

    Gallery:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/GalleryItem"
      required: [id, userId, items]

    UpdateGalleryBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/GalleryItem"

    RequestUploadUrlBody:
      type: object
      properties:
        name:
          type: string
        size:
          type: integer
        contentType:
          type: string
      required: [name, size, contentType]

    RequestUploadUrlResponse:
      type: object
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
      required: [uploadURL, objectPath]