# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create account
 */
export const RegisterBody = zod.object({
  email: zod.string(),
  password: zod.string(),
  name: zod.string(),
  username: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  username: zod.string(),
});

/**
 * @summary Get current session
 */
export const GetMeResponse = zod.object({
  user: zod.union([
    zod.object({
      id: zod.number(),
      email: zod.string(),
      name: zod.string(),
      username: zod.string(),
    }),
    zod.null(),
  ]),
});

/**
 * @summary Get my landing page
 */
export const GetMyLandingResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  template: zod.string(),
  title: zod.string(),
  subtitle: zod.string().nullish(),
  description: zod.string().nullish(),
  appFileObjectPath: zod.string().nullish(),
  tutorialVideoObjectPath: zod.string().nullish(),
  tutorialVideoUrl: zod.string().nullish(),
  buttonText: zod.string().nullish(),
  photos: zod.array(
    zod.object({
      objectPath: zod.string(),
      caption: zod.string().nullish(),
    }),
  ),
  socialLinks: zod.array(
    zod.object({
      platform: zod.string(),
      url: zod.string(),
      label: zod.string(),
    }),
  ),
});

/**
 * @summary Update my landing page
 */
export const UpdateMyLandingBody = zod.object({
  template: zod.string().optional(),
  title: zod.string().optional(),
  subtitle: zod.string().nullish(),
  description: zod.string().nullish(),
  appFileObjectPath: zod.string().nullish(),
  tutorialVideoObjectPath: zod.string().nullish(),
  tutorialVideoUrl: zod.string().nullish(),
  buttonText: zod.string().nullish(),
  photos: zod
    .array(
      zod.object({
        objectPath: zod.string(),
        caption: zod.string().nullish(),
      }),
    )
    .optional(),
  socialLinks: zod
    .array(
      zod.object({
        platform: zod.string(),
        url: zod.string(),
        label: zod.string(),
      }),
    )
    .optional(),
});

export const UpdateMyLandingResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  template: zod.string(),
  title: zod.string(),
  subtitle: zod.string().nullish(),
  description: zod.string().nullish(),
  appFileObjectPath: zod.string().nullish(),
  tutorialVideoObjectPath: zod.string().nullish(),
  tutorialVideoUrl: zod.string().nullish(),
  buttonText: zod.string().nullish(),
  photos: zod.array(
    zod.object({
      objectPath: zod.string(),
      caption: zod.string().nullish(),
    }),
  ),
  socialLinks: zod.array(
    zod.object({
      platform: zod.string(),
      url: zod.string(),
      label: zod.string(),
    }),
  ),
});

/**
 * @summary Get my gallery
 */
export const GetMyGalleryResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string().nullish(),
  description: zod.string().nullish(),
  items: zod.array(
    zod.object({
      id: zod.string(),
      objectPath: zod.string(),
      type: zod.enum(["photo", "video"]),
      caption: zod.string().nullish(),
      order: zod.number(),
    }),
  ),
});

/**
 * @summary Update my gallery
 */
export const UpdateMyGalleryBody = zod.object({
  title: zod.string().nullish(),
  description: zod.string().nullish(),
  items: zod
    .array(
      zod.object({
        id: zod.string(),
        objectPath: zod.string(),
        type: zod.enum(["photo", "video"]),
        caption: zod.string().nullish(),
        order: zod.number(),
      }),
    )
    .optional(),
});

export const UpdateMyGalleryResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string().nullish(),
  description: zod.string().nullish(),
  items: zod.array(
    zod.object({
      id: zod.string(),
      objectPath: zod.string(),
      type: zod.enum(["photo", "video"]),
      caption: zod.string().nullish(),
      order: zod.number(),
    }),
  ),
});

/**
 * @summary Get public landing page by username
 */
export const GetPublicLandingParams = zod.object({
  username: zod.coerce.string(),
});

export const GetPublicLandingResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  template: zod.string(),
  title: zod.string(),
  subtitle: zod.string().nullish(),
  description: zod.string().nullish(),
  appFileObjectPath: zod.string().nullish(),
  tutorialVideoObjectPath: zod.string().nullish(),
  tutorialVideoUrl: zod.string().nullish(),
  buttonText: zod.string().nullish(),
  photos: zod.array(
    zod.object({
      objectPath: zod.string(),
      caption: zod.string().nullish(),
    }),
  ),
  socialLinks: zod.array(
    zod.object({
      platform: zod.string(),
      url: zod.string(),
      label: zod.string(),
    }),
  ),
});

/**
 * @summary Get public gallery by username
 */
export const GetPublicGalleryParams = zod.object({
  username: zod.coerce.string(),
});

export const GetPublicGalleryResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  title: zod.string().nullish(),
  description: zod.string().nullish(),
  items: zod.array(
    zod.object({
      id: zod.string(),
      objectPath: zod.string(),
      type: zod.enum(["photo", "video"]),
      caption: zod.string().nullish(),
      order: zod.number(),
    }),
  ),
});

/**
 * @summary Request presigned upload URL
 */
export const RequestUploadUrlBody = zod.object({
  name: zod.string(),
  size: zod.number(),
  contentType: zod.string(),
});

export const RequestUploadUrlResponse = zod.object({
  uploadURL: zod.string(),
  objectPath: zod.string(),
});