# lib/api-zod/src/generated/types/gallery.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GalleryItem } from "./galleryItem";

export interface Gallery {
  id: number;
  userId: number;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  description?: string | null;
  items: GalleryItem[];
}