# lib/api-zod/src/generated/types/updateLandingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LandingPhoto } from "./landingPhoto";
import type { SocialLink } from "./socialLink";

export interface UpdateLandingBody {
  template?: string;
  title?: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  appFileObjectPath?: string | null;
  /** @nullable */
  tutorialVideoObjectPath?: string | null;
  /** @nullable */
  tutorialVideoUrl?: string | null;
  /** @nullable */
  buttonText?: string | null;
  photos?: LandingPhoto[];
  socialLinks?: SocialLink[];
}